implemented some UI changes in loginpage
