+++ lib/authentication/secondPage.dart
@@ -1,6 +1,7 @@
 import 'package:Runbhumi/authentication/loginPage.dart';
 import 'package:Runbhumi/authentication/signUp.dart';
 import 'package:Runbhumi/components/button.dart';
+import 'package:Runbhumi/components/googleOauth.dart';
 import 'package:flutter/material.dart';
 
 class SecondPage extends StatefulWidget {
@@ -34,7 +35,7 @@ class _SecondPageState extends State<SecondPage> {
               myWidget: SignUp(),
             ),
             DividingOr(),
-            GoogleOauthBig(),
+            GoogleOauth(),
             SizedBox(
               height: 20,
             ),
@@ -42,57 +43,6 @@ class _SecondPageState extends State<SecondPage> {
         ),
       ),
     );
-  }
-}
-
-class GoogleOauthBig extends StatelessWidget {
-  const GoogleOauthBig({
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-        onTap: () {},
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16.0),
-              child: Container(
-                height: 55,
-                decoration: BoxDecoration(
-                    color: Theme.of(context).buttonColor,
-                    borderRadius: BorderRadius.circular(8),
-                    boxShadow: [
-                      BoxShadow(
-                          color: Theme.of(context).buttonColor.withOpacity(.4),
-                          blurRadius: 6,
-                          offset: Offset(0, 5))
-                    ]),
-                child: Center(
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: <Widget>[
-                      Padding(
-                        padding: const EdgeInsets.only(left: 16.0),
-                        child:
-                            Image(image: AssetImage('assets/googleicon.png')),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                        child: Text(
-                          "Continue with Google",
-                          style: TextStyle(color: Colors.black, fontSize: 20),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ));
   }
 }
