+++ test/widget_test.dart
@@ -5,7 +5,7 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:Runbhumi/authentication/secondPage.dart';
+import 'package:Runbhumi/components/googleOauth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -30,7 +30,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   MaterialApp app = MaterialApp(
-    home: Scaffold(body: GoogleOauthBig()),
+    home: Scaffold(body: GoogleOauth()),
   );
   testWidgets('GoogleOauthBig UI Test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
